Move announcement helper and add constants
Refactor announcement code by moving the check_over_800 helper into commands/helper.py and centralizing related constants in constants.py (MESSAGELINKPREFIX, MAX_CHAR_LIMIT, TRUNCATE_OFFSET).
Simplify the paginator status_title logic.
Minor comment and formatting cleanups included.
