+++ services/moddingway/moddingway/constants.py
@@ -76,3 +76,8 @@ MAX_APPEAL_REASON_LENGTH = 1024
 MAX_SEND_MESSAGE_LENGTH = 250
 MAX_SLOWMODE_INTERVAL = 21600  # 6 hours in seconds
 PERMANENT_BAN_STRIKE_THRESHOLD = 15
+
+MESSAGELINKPREFIX = "https://discord.com/channels/"
+
+MAX_CHAR_LIMIT = 800
+TRUNCATE_OFFSET = 3
